modificacion de Ubicacion

+++ app/src/main/java/unah/proyecto/aeo/aplicacionagendaelectronicaoriental/EntidadesBD/Contactos.java
@@ -13,10 +13,10 @@ public class Contactos {
     private int imagen;
     private String e_mail;
     private String descripcion_organizacion;
-    private int latitud;
-    private int longitud;
+    private double latitud;
+    private double longitud;
 
-    public Contactos(int id_contacto, String nombre_organizacion, String numero_fijo, String numero_movil, String direccion, int imagen, String e_mail, String descripcion_organizacion, int latitud, int longitud) {
+    public Contactos(int id_contacto, String nombre_organizacion, String numero_fijo, String numero_movil, String direccion, int imagen, String e_mail, String descripcion_organizacion, double latitud, double longitud) {
         this.id_contacto = id_contacto;
         this.nombre_organizacion = nombre_organizacion;
         this.numero_fijo = numero_fijo;
@@ -93,7 +93,7 @@ public class Contactos {
         this.descripcion_organizacion = descripcion_organizacion;
     }
 
-    public int getLatitud() {
+    public double getLatitud() {
         return latitud;
     }
 
@@ -101,7 +101,7 @@ public class Contactos {
         this.latitud = latitud;
     }
 
-    public int getLongitud() {
+    public double getLongitud() {
         return longitud;
     }
